less boilerplate in DAOs #87: user skinny-orm for roundDao

+++ app/org/intracer/wmua/Round.scala
@@ -6,7 +6,7 @@ import org.joda.time.DateTime
 import org.scalawiki.dto.Page
 
 case class Round(id: Option[Long],
-                 number: Int,
+                 number: Long,
                  name: Option[String] = None,
                  contest: Long,
                  roles: Set[String] = Set("jury"),
